requirement.txt
django==2.2.5
djangorestframework==3.10.3
Markdown==3.1.1
django-filter==2.2.0
django-cors-headers==3.1.0
mysqlclient==1.4.4
Pillow==6.1.0
requests==2.22.0
django-crispy-forms==1.7.2
django-import-export==1.2.0
django-reversion==3.0.4
django-formtools==2.1
future==0.17.1
httplib2==0.13.1
six==1.12.0
xlwt==1.3.0
xlsxwriter==1.2.0
coreapi==2.3.3